重构: 使用PlatformIO多环境配置分离测试和生产代码
- 修改 platformio.ini 添加测试环境配置
- 创建 src/test_main.cpp 专门用于测试环境
- 简化 src/main.cpp 为生产环境代码 (从514行减少到118行)
- 生产环境排除测试文件，测试环境排除生产入口文件
- 两个环境都编译测试通过

使用方法:
- 生产环境: pio run -e esp32-s3-zero
- 测试环境: pio run -e test
