+++ platformio.ini
@@ -8,6 +8,7 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+; 生产环境配置
 [env:esp32-s3-zero]
 platform = espressif32
 board    = esp32-s3-devkitc-1     ; 借用的板型
@@ -28,5 +29,30 @@ monitor_speed = 115200
 lib_deps =
     bblanchon/ArduinoJson@^6.21.3
 
-; --- 使用统一的测试架构，不再需要排除文件 ---
-build_src_filter = +<*>
+; --- 生产环境：排除测试入口文件 ---
+build_src_filter = +<*> -<test_main.cpp>
+
+; 测试环境配置
+[env:test]
+platform = espressif32
+board    = esp32-s3-devkitc-1     ; 借用的板型
+framework = arduino
+
+; --- 让 USB-CDC 作为默认串口 ---
+build_flags =
+  -DARDUINO_USB_CDC_ON_BOOT=1
+  -DBOARD_HAS_PSRAM              ; 启用 2 MB PSRAM
+  -DENABLE_TESTING=1             ; 启用测试模式标志
+
+; --- 真实硬件修正 ---
+board_upload.flash_size = 4MB
+board_build.partitions  = default.csv   ; 4 MB 分区表
+
+monitor_speed = 115200
+
+; --- 依赖库 ---
+lib_deps =
+    bblanchon/ArduinoJson@^6.21.3
+
+; --- 测试环境：使用测试入口文件，排除生产入口文件 ---
+build_src_filter = +<*> -<main.cpp>
